Down the rabbit hole the wrong way, bay to M in EntityEncoder/Decoder
