--- core/src/main/scala/org/http4s/Req.scala
@@ -1,5 +1,0 @@
-package org.http4s
-
-class Req {
-
-}
