--- core/src/main/scala/org/http4s/Resp.scala
@@ -1,5 +1,0 @@
-package org.http4s
-
-class Resp {
-
-}
