+++ core/src/main/scala/org/http4s/syntax/AllSyntax.scala
@@ -4,8 +4,6 @@ package syntax
 trait AllSyntax
     extends AnyRef
     with AsyncSyntax
-    with EffectResponseSyntax
-    with EffectRequestSyntax
     with KleisliSyntax
     with NonEmptyListSyntax
     with StringSyntax
